refactor(frontend): move drag/drop/delete canvas handlers into useCanvasDragDrop hook
